Migrate build system from GCC to Clang with strict clang-tidy enforcement
This commit migrates the entire project from GCC to Clang compiler with
comprehensive clang-tidy static analysis and strict warnings-as-errors.

Key changes:

Build Configuration:
- Add CMakePresets.json with clang-debug and clang-release configurations
- Update .clang-tidy: disable misc-include-cleaner, update HeaderFilterRegex

Code Quality Improvements (200+ fixes across 30+ files):
- Const-correctness: 100+ variables marked const for immutability
- Narrowing conversions: explicit static_cast for all type conversions
- Modern C++: pass-by-value with std::move() for string parameters
- Type aliases: replace typedef with using declarations
- Returns: use braced-init-list instead of explicit type construction
- Initialization: move member initialization to initializer lists
- Portability: replace pragma once with traditional include guards

Platform-Specific Handling:
- Add NOLINT suppressions for QMutexLocker (destructor side-effects)
- Add NOLINT for PC/SC C-array requirements (BYTE response[4096])
- Add NOLINT for X11 API requirements (char keys[32])
- Add NOLINT for Qt moc compilation patterns

Modified Components:
- Daemon: D-Bus objects, OATH protocol, device manager, workflows
- KRunner: plugin, match builder, configuration
- Config: KCM module
- Shared: D-Bus proxies, UI dialogs, value types
- Tests: display strategies, modifier key checker

Build Results:
- All binaries build successfully with 0 errors
- yubikey-oath-daemon: 23MB (debug build)
- krunner_yubikey.so: 7.4MB
- kcm_krunner_yubikey.so: 7.1MB
- All test executables built

Benefits:
- Improved type safety and const-correctness
- Better static analysis catching potential bugs
- More portable code (include guards)
- Modern C++ patterns throughout codebase
- Stricter compile-time checks

+++ src/daemon/formatting/code_validator.cpp
@@ -13,14 +13,14 @@ int CodeValidator::calculateCodeValidity()
     // TOTP codes are valid for 30 seconds
     // Calculate remaining seconds in current 30-second window
     qint64 const currentTime = QDateTime::currentSecsSinceEpoch();
-    return TOTP_PERIOD - (currentTime % TOTP_PERIOD);
+    return static_cast<int>(TOTP_PERIOD - (currentTime % TOTP_PERIOD));
 }
 
 QDateTime CodeValidator::calculateExpirationTime(const QDateTime &currentTime)
 {
     // Calculate remaining seconds based on provided currentTime, not actual current time
     qint64 const timeInSeconds = currentTime.toSecsSinceEpoch();
-    int const remainingSeconds = TOTP_PERIOD - (timeInSeconds % TOTP_PERIOD);
+    int const remainingSeconds = static_cast<int>(TOTP_PERIOD - (timeInSeconds % TOTP_PERIOD));
     return currentTime.addSecs(remainingSeconds);
 }
